requirements.txt
APScheduler==2.1.2
Mako==1.0.7
ansicolors==1.0.2
future==0.17.1
futures==2.2.0
motor==2.0.0
pymongo==3.7.1
pystache==0.5.3
python-dateutil==2.5.3; python_version < "3.7"
python-dateutil==2.8.2; python_version >= "3.7"
redis==2.8.0
requests==2.28.0; python_version >= "3"
requests[security]==2.24.0; python_version < "3"
setuptools==30.0.0
simplejson==3.5.2
tornado-redis==2.4.16
tornado==4.4.1
toro==0.5
typing==3.7.4.1; python_version < "3"
